+++ include/LIEF/MachO/Binary.hpp
@@ -83,9 +83,10 @@ class LIEF_API Binary : public LIEF::Binary  {
   //! If a symbol with the given name exists
   bool has_symbol(const std::string& name) const;
 
-  //! Return Symbol from the given name
-  const Symbol& get_symbol(const std::string& name) const;
-  Symbol& get_symbol(const std::string& name);
+  //! Return Symbol from the given name. If the symbol does not
+  //! exists, it returns a null pointer
+  const Symbol* get_symbol(const std::string& name) const;
+  Symbol* get_symbol(const std::string& name);
 
   //! Check if the given symbol is an exported one
   static bool is_exported(const Symbol& symbol);