+++ src/MachO/BinaryParser.cpp
@@ -137,10 +137,10 @@ void BinaryParser::parse_export_trie(uint64_t start, uint64_t end, const std::st
 
     const std::string& symbol_name = prefix;
     std::unique_ptr<ExportInfo> export_info{new ExportInfo{0, flags, offset}};
-    if (this->binary_->has_symbol(symbol_name)) {
-      Symbol& symbol = this->binary_->get_symbol(symbol_name);
-      export_info->symbol_ = &symbol;
-      symbol.export_info_ = export_info.get();
+    Symbol* symbol = this->binary_->get_symbol(symbol_name);
+    if (symbol != nullptr) {
+      export_info->symbol_ = symbol;
+      symbol->export_info_ = export_info.get();
     } else { // Register it into the symbol table
       std::unique_ptr<Symbol> symbol{new Symbol{}};
       symbol->origin_            = SYMBOL_ORIGINS::SYM_ORIGIN_DYLD_EXPORT;
@@ -167,10 +167,10 @@ void BinaryParser::parse_export_trie(uint64_t start, uint64_t end, const std::st
         imported_name = export_info->symbol().name();
       }
 
-      if (this->binary_->has_symbol(imported_name)) {
-        Symbol& symbol = this->binary_->get_symbol(imported_name);
-        export_info->alias_ = &symbol;
-        symbol.export_info_ = export_info.get();
+      Symbol* symbol = this->binary_->get_symbol(imported_name);
+      if (symbol != nullptr) {
+        export_info->alias_ = symbol;
+        symbol->export_info_ = export_info.get();
       } else {
         std::unique_ptr<Symbol> symbol{new Symbol{}};
         symbol->origin_            = SYMBOL_ORIGINS::SYM_ORIGIN_DYLD_EXPORT;